requirements.txt
google-generativeai>=0.3.0
pymongo>=4.6.0
pypdf>=3.17.0
python-dotenv>=1.0.0
openai>=1.12.0
